backend/pyproject.toml
[project]
name = "ctrl-contabil-backend"
version = "0.1.0"
description = "FastAPI backend for Ctrl Contabil with Authentik OIDC integration."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "PyJWT[crypto]>=2.10.0,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "SQLAlchemy[asyncio]>=2.0.30,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0"
]

[project.optional-dependencies]
saml = ["python3-saml>=1.16.0,<2.0.0"]
test = ["mypy>=1.11.0,<2.0.0", "pytest>=8.3.0,<9.0.0", "ruff>=0.6.0,<1.0.0"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S"]
ignore = ["S105", "S106"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
